mm: consider killable task's anon as reclaimable
zone_reclaimable_pages which is used by should_reclaim_retry to
make a decision on retrying reclaim does not consider anon pages
if swap is full. This logic works on systems which rely on OOM
to kill tasks and free up memory. But when reclaimers like
android lowmemorykiller is used, reclaim should retry enough so
that lowmemorykiller is invoked via shrinkers, thus avoiding OOM.
This fixes the premature order 0 OOM killer invocations observed
when swap is full.

Change-Id: I99e6585b5a5dd2d871678c13349e7a62d5aefee0
Signed-off-by: Vinayak Menon <vinmenon@codeaurora.org>
Signed-off-by: atndko <z1281552865@gmail.com>